Item Lists Finished
Finished creating the item lists, and confirmed proper implementation in the program.cs. The ideas i could do with this are starting to get a little wild, would like to learn automation through python to script this to the game in order to do the math on on the client in real time instead of having to manually input numbers. The plan forward is to finish this in a console application in order to essentially create the offline wikipedia FIRST, and then worry about automation. Tomorrow I will work on implementing the descriptions and math behind each specific item.